Added Forum Controllers and Views

+++ Education.BLL/Services/ForumServices/MessageService.cs
@@ -1,4 +1,5 @@
 ﻿using Education.BLL.DTO.Forum;
+using Education.BLL.DTO.Forum.Edit;
 using Education.BLL.DTO.User;
 using Education.BLL.Logic.Interfaces;
 using Education.BLL.Services.ForumServices.Interfaces;
@@ -27,7 +28,7 @@ namespace Education.BLL.Services.ForumServices
             MessageRules = rules;
         }
 
-        private void EditMessage(Message message, MessageDTO messageDTO, User user, bool edit = true)
+        private void EditMessage(Message message, MessageEditDTO messageDTO, User user, bool edit = true)
         {
             message.Text = messageDTO.Text;
             if (edit)
@@ -42,6 +43,13 @@ namespace Education.BLL.Services.ForumServices
             }
         }
 
+        private bool IsFirstMessage(Message message)
+        {
+            if (message.Theme == null) return false;
+            if (message?.Theme?.Messages?.OrderBy(X => X.Id).FirstOrDefault() == message) return true;
+            return false;
+        }
+
         private (AccessCode Code, Message Message, User User) CheckMessage(UserDTO userDTO, int MessageId, Func<User, Message, bool> checkFunc, IUOW Data)
         {
             var user = GetUserService.Get(userDTO, Data);
@@ -51,7 +59,7 @@ namespace Education.BLL.Services.ForumServices
             else return (AccessCode.NoPremision, null, user);
         }
 
-        public CreateResultDTO Create(MessageDTO DTO, UserDTO userDTO)
+        public CreateResultDTO Create(MessageEditDTO DTO, UserDTO userDTO)
         {
             using(var Data = DataFactory.Get())
             {
@@ -62,6 +70,7 @@ namespace Education.BLL.Services.ForumServices
                 {
                     var message = new Message();
                     EditMessage(message, DTO, user, false);
+                    message.Theme = theme;
                     Data.MessageRepository.Add(message);
                     Data.SaveChanges();
                     return new CreateResultDTO(message.Id, AccessCode.Succsess);
@@ -77,6 +86,7 @@ namespace Education.BLL.Services.ForumServices
                 var check = CheckMessage(userDTO, id, MessageRules.CanDelete, Data);
                 if (check.Code == AccessCode.Succsess)
                 {
+                    if (IsFirstMessage(check.Message)) return AccessCode.NoPremision;
                     Data.MessageRepository.Delete(check.Message);
                     Data.SaveChanges();
                 }
@@ -95,18 +105,30 @@ namespace Education.BLL.Services.ForumServices
             }
         }
 
-        public AccessCode Update(MessageDTO DTO, UserDTO userDTO)
+        public AccessCode Update(MessageEditDTO DTO, UserDTO userDTO)
         {
             using (var Data = DataFactory.Get())
             {
                 var check = CheckMessage(userDTO, DTO.Id, MessageRules.CanRead, Data);
                 if (check.Code == AccessCode.Succsess)
                 {
+                    if (IsFirstMessage(check.Message)) return AccessCode.NoPremision;
                     EditMessage(check.Message, DTO, check.User);
                     Data.MessageRepository.Edited(check.Message);
                     Data.SaveChanges();
                 }
                 return check.Code;
+            }
+        }
+
+        public bool CanCreate(int ThemeId, UserDTO userDTO)
+        {
+            using (var Data = DataFactory.Get())
+            {
+                var user = GetUserService.Get(userDTO, Data);
+                var theme = Data.ThemeRepository.Get().FirstOrDefault(x => x.Id == ThemeId);
+                if (theme == null) return false;
+                return MessageRules.CanCreate(user, theme);
             }
         }
     }
